#267 Prepare JavaDsl support
