+++ tools/test-generator/src/main/java/com/consol/citrus/generate/javadsl/JavaDslTestGenerator.java
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.generate;
+package com.consol.citrus.generate.javadsl;
+
+import com.consol.citrus.generate.UnitFramework;
 
 import java.util.Properties;
