+++ tools/test-generator/src/main/java/com/consol/citrus/generate/xml/WsdlXmlTestGenerator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.generate;
+package com.consol.citrus.generate.xml;
 
 import com.consol.citrus.context.TestContext;
 import com.consol.citrus.exceptions.CitrusRuntimeException;
