+++ tools/test-generator/src/test/java/com/consol/citrus/generate/xml/MessagingXmlTestGeneratorTest.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.generate;
+package com.consol.citrus.generate.xml;
 
 import com.consol.citrus.Citrus;
+import com.consol.citrus.generate.UnitFramework;
 import com.consol.citrus.message.DefaultMessage;
 import com.consol.citrus.util.FileUtils;
 import org.springframework.core.io.FileSystemResource;
