+++ tools/test-generator/src/test/java/com/consol/citrus/generate/xml/SwaggerXmlTestGeneratorTest.java
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.generate;
+package com.consol.citrus.generate.xml;
 
 import com.consol.citrus.Citrus;
+import com.consol.citrus.generate.TestGenerator;
+import com.consol.citrus.generate.UnitFramework;
 import com.consol.citrus.util.FileUtils;
 import org.springframework.core.io.FileSystemResource;
 import org.testng.Assert;
@@ -29,27 +31,49 @@ import java.io.IOException;
  * @author Christoph Deppisch
  * @since 2.7.4
  */
-public class WsdlXmlTestGeneratorTest {
+public class SwaggerXmlTestGeneratorTest {
 
     @Test
-    public void testCreateTest() throws IOException {
-        WsdlXmlTestGenerator generator = new WsdlXmlTestGenerator();
+    public void testCreateTestAsClient() throws IOException {
+        SwaggerXmlTestGenerator generator = new SwaggerXmlTestGenerator();
 
         generator.withAuthor("Christoph")
                 .withDescription("This is a sample test")
                 .usePackage("com.consol.citrus")
                 .withFramework(UnitFramework.TESTNG);
 
-        generator.withWsdl("com/consol/citrus/wsdl/BookStore.wsdl");
+        generator.withNamePrefix("UserLoginClient_");
+        generator.withSpec("com/consol/citrus/swagger/user-login-api.json");
 
         generator.create();
 
-        verifyTest("BookStore_addBook_IT", "book:addBook", "book:addBookResponse");
-        verifyTest("BookStore_addBookAudio_IT", "aud:addBookAudio", "aud:addBookAudioResponse");
-        verifyTest("BookStore_deleteBook_IT", "book:deleteBook", "book:deleteBookResponse");
+        verifyTest("UserLoginClient_createUser_IT");
+        verifyTest("UserLoginClient_loginUser_IT");
+        verifyTest("UserLoginClient_logoutUser_IT");
+        verifyTest("UserLoginClient_getUserByName_IT");
     }
 
-    private void verifyTest(String name, String requestName, String responseName) throws IOException {
+    @Test
+    public void testCreateTestAsServer() throws IOException {
+        SwaggerXmlTestGenerator generator = new SwaggerXmlTestGenerator();
+
+        generator.withAuthor("Christoph")
+                .withDescription("This is a sample test")
+                .usePackage("com.consol.citrus")
+                .withFramework(UnitFramework.TESTNG);
+
+        generator.withMode(TestGenerator.GeneratorMode.SERVER);
+        generator.withSpec("com/consol/citrus/swagger/user-login-api.json");
+
+        generator.create();
+
+        verifyTest("UserLoginService_createUser_IT");
+        verifyTest("UserLoginService_loginUser_IT");
+        verifyTest("UserLoginService_logoutUser_IT");
+        verifyTest("UserLoginService_getUserByName_IT");
+    }
+
+    private void verifyTest(String name) throws IOException {
         File javaFile = new File(Citrus.DEFAULT_TEST_SRC_DIRECTORY + "java/com/consol/citrus/" + name + ".java");
         Assert.assertTrue(javaFile.exists());
 
@@ -67,8 +91,6 @@ public class WsdlXmlTestGeneratorTest {
         Assert.assertTrue(xmlContent.contains("<author>Christoph</author>"));
         Assert.assertTrue(xmlContent.contains("<description>This is a sample test</description>"));
         Assert.assertTrue(xmlContent.contains("<testcase name=\"" + name + "\">"));
-        Assert.assertTrue(xmlContent.contains("<data>&lt;" + requestName));
-        Assert.assertTrue(xmlContent.contains("<data>&lt;" + responseName));
     }
 
 }
